update schools migration to match data

+++ db/migrate/20170304195222_create_schools.rb
@@ -1,12 +1,12 @@
 class CreateSchools < ActiveRecord::Migration[5.0]
   def change
 		create_table :schools do |t|
-			t.float :objectid
+			t.integer :objectid
 			t.float :gis_nr_sde_schools_area
 			t.float :perimeter
-			t.float :poi_
-			t.float :poi_id
-			t.float :symbol
+			t.integer :poi_
+			t.integer :poi_id
+			t.integer :symbol
 			t.string :school
 			t.string :address
 			t.string :type_of_school
